Done some time conversation and TimePickerDialog

+++ app/src/main/java/ch/hsr/sa/radiotour/dataaccess/models/RaceGroup.java
@@ -1,5 +1,7 @@
 package ch.hsr.sa.radiotour.dataaccess.models;
 
+import java.util.Date;
+
 import io.realm.RealmList;
 import io.realm.RealmObject;
 import io.realm.annotations.PrimaryKey;
@@ -10,8 +12,8 @@ public class RaceGroup extends RealmObject {
     private String id;
     @Required
     private String type;
-    private int actualGapTime;
-    private int historyGapTime;
+    private long actualGapTime;
+    private long historyGapTime;
     private int position;
 
     private RealmList<Rider> riders;
@@ -44,19 +46,19 @@ public class RaceGroup extends RealmObject {
         this.type = type.toString();
     }
 
-    public int getActualGapTime() {
+    public long getActualGapTime() {
         return actualGapTime;
     }
 
-    public void setActualGapTime(int actualGapTime) {
+    public void setActualGapTime(long actualGapTime) {
         this.actualGapTime = actualGapTime;
     }
 
-    public int getHistoryGapTime() {
+    public long getHistoryGapTime() {
         return historyGapTime;
     }
 
-    public void setHistoryGapTime(int historyGapTime) {
+    public void setHistoryGapTime(long historyGapTime) {
         this.historyGapTime = historyGapTime;
     }
